Revert "not working order service"
This reverts commit bdf570df276ae5079e34fe45c797ccce13205e75.

+++ OrderService.Infrastructure/Repository/OrderRepository.cs
@@ -9,23 +9,23 @@ namespace OrderService.Infrastructure.Repository;
 
 public class OrderRepository : IOrderRepository
 {
-    private readonly IMongoCollection<Order> _orders;
+    private readonly IMongoService _mongoService;
 
     public OrderRepository(IMongoService mongoService)
     {
-        _orders = mongoService.Orders;
+        _mongoService = mongoService;
     }
 
 
     public async Task CreateAsync(Order newOrder)
     {
-        await _orders.InsertOneAsync(newOrder);
+        await _mongoService.Orders.InsertOneAsync(newOrder);
     }
 
 
     public async Task UpdateAsync(Order updatedOrder)
     {
-        var result = await _orders.ReplaceOneAsync(o => o.Id == updatedOrder.Id, updatedOrder);
+        var result = await _mongoService.Orders.ReplaceOneAsync(o => o.Id == updatedOrder.Id, updatedOrder);
         if (!result.IsModifiedCountAvailable && result.ModifiedCount == 0)
             throw new NotFoundException<Order>(updatedOrder.Id);
     }
@@ -33,15 +33,15 @@ public class OrderRepository : IOrderRepository
 
     public async Task DeleteAsync(string orderId)
     {
-        var result = await _orders.DeleteOneAsync(o => o.Id == orderId);
+        var result = await _mongoService.Orders.DeleteOneAsync(o => o.Id == orderId);
         if (result.DeletedCount == 0)
             throw new NotFoundException<Order>(orderId);
     }
 
 
-    public async Task<PagedList<Order>> GetOrdersPaged(RequestParameters requestParameters)
+    public async Task<PagedList<Order>> GetAll(RequestParameters requestParameters)
     {
-        var orders = await _orders
+        var orders = await _mongoService.Orders
             .Search(requestParameters.SearchTerm)
             .CustomSort(requestParameters.OrderBy)
             .ToListAsync();
@@ -55,7 +55,8 @@ public class OrderRepository : IOrderRepository
 
     public async Task<Order> GetWithId(string orderId)
     {
-        var order = await _orders.Find(o => o.Id == orderId).FirstOrDefaultAsync();
+        var order = await _mongoService.Orders.Find(o => o.Id == orderId).FirstOrDefaultAsync();
+        var x = new FormatException();
         if (order == null)
             throw new NotFoundException<Order>(orderId);
         return order;
@@ -64,7 +65,7 @@ public class OrderRepository : IOrderRepository
 
     public async Task ChangeStatus(string orderId, string newStatus)
     {
-        var result = await _orders.UpdateOneAsync(
+        var result = await _mongoService.Orders.UpdateOneAsync(
             o => o.Id == orderId,
             Builders<Order>.Update
                 .Set(o => o.Status, newStatus));
@@ -76,7 +77,7 @@ public class OrderRepository : IOrderRepository
 
     public async Task<PagedList<Order>> GetOrdersOfCustomer(string customerId, RequestParameters requestParameters)
     {
-        var orders = await _orders
+        var orders = await _mongoService.Orders
             .Search(requestParameters.SearchTerm, customerId)
             .CustomSort(requestParameters.OrderBy)
             .ToListAsync();
@@ -89,18 +90,22 @@ public class OrderRepository : IOrderRepository
 
     public async Task<bool> DeleteOrderOfCustomer(string customerId)
     {
-        var result = await _orders.DeleteManyAsync(o => o.CustomerId == customerId);
+        var result = await _mongoService.Orders.DeleteManyAsync(o => o.CustomerId == customerId);
         return result.DeletedCount > 0;
     }
 
-    public async Task UpdateProductRelatedOrders(Product product)
+    public async Task UpdateProductRelatedOrders(string productId)
     {
-        
-        var builder = Builders<Order>.Filter;
-        var filter = builder.AnyEq(o => o.ProductIds, product.Id);
-        await _orders.Find(filter).ForEachAsync(async order =>
+        var product =await _mongoService.Products.Find(p => p.Id == productId).FirstOrDefaultAsync();
+        if (product == null)
         {
-            order.ProductIds.Remove(product.Id);
+            throw new NotFoundException<Product>(productId);
+        }
+        var builder = Builders<Order>.Filter;
+        var filter = builder.AnyEq(o => o.ProductIds, productId);
+        await _mongoService.Orders.Find(filter).ForEachAsync(async order =>
+        {
+            order.ProductIds.Remove(productId);
             if (order.ProductIds.Count == 0)
                 await DeleteAsync(order.Id);
             else
